backend/requirements.txt
django==3.2.15
djangorestframework==3.13.1
python-dotenv==0.20.0
pytest==7.1.3
django-mptt==0.13.4
psycopg2-binary==2.9.3
gunicorn==20.1.0
django-debug-toolbar==3.6.0
djangorestframework-recursive==0.1.2